WIP: Move setting toolbar max-width to inlineeditoruiview
